+++ app/scripts/controllers/transactions/bitbox-utils.js
@@ -1,4 +1,4 @@
-const SLPSDK = require('slp-sdk/lib/SLP').default
+const SLPSDK = require('slp-sdk')
 const SLP = new SLPSDK()
 const BigNumber = require('slpjs/node_modules/bignumber.js')
 const slpjs = require('slpjs')
@@ -73,21 +73,23 @@ class BitboxUtils {
   static async publishTx (hex) {
     return new Promise((resolve, reject) => {
       SLP.RawTransactions.sendRawTransaction(hex).then(
-        // TODO: pass back result instead of result[0] after we update REST to return string instead of array
         result => {
           try {
-            if (result[0].length !== 64) {
-              // TODO: Validate result is a txid
-              reject('Transaction failed: ' + result)
+            if (result.length !== 64) {
+              reject(result)
             } else {
-              resolve(result[0])
+              resolve(result)
             }
           } catch (ex) {
             reject(ex)
           }
         },
         err => {
-          reject(err)
+          if (err.error) {
+            reject(err.error)
+          } else {
+            reject(err)
+          }
         }
       )
     })
@@ -127,6 +129,13 @@ class BitboxUtils {
         })
 
         const satoshisRemaining = totalUtxoAmount - byteCount - satoshisToSend
+
+        // Verify sufficient fee
+        if (satoshisRemaining < 0) {
+          throw new Error(
+            'Not enough Bitcoin Cash for fee. Deposit a small amount and try again.'
+          )
+        }
 
         // Destination output
         transactionBuilder.addOutput(to, satoshisToSend)
@@ -182,6 +191,12 @@ class BitboxUtils {
         ).decimalPlaces(tokenDecimals)
         const tokenSendAmount = scaledTokenSendAmount.times(10 ** tokenDecimals)
 
+        if (tokenSendAmount.lt(1)) {
+          throw new Error(
+            'Amount below minimum for this token. Increase the send amount and try again.'
+          )
+        }
+
         let tokenBalance = new BigNumber(0)
         for (const tokenUtxo of spendableTokenUtxos) {
           const utxoBalance = tokenUtxo.slp.quantity
@@ -219,6 +234,13 @@ class BitboxUtils {
         )
 
         const satoshisRemaining = totalUtxoAmount - byteCount
+
+        // Verify sufficient fee
+        if (satoshisRemaining < 0) {
+          throw new Error(
+            'Not enough Bitcoin Cash for fee. Deposit a small amount and try again.'
+          )
+        }
 
         // SLP data output
         transactionBuilder.addOutput(sendOpReturn, 0)