+++ ui/app/components/identicon.js
@@ -57,6 +57,7 @@ IdenticonComponent.prototype.render = function () {
   } else {
     return h('img.balance-icon', {
       className,
+      // src: '../../../../node_modules/bch-token-icons/svg/icon/bch.svg',
       src: './images/bch_logo.svg',
       style: {
         ...style,