+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Current Develop Branch
 
+## 0.3.1
+-Improve error messages
+
+## 0.3.0
+-Show recent BCH send and receive transaction history
+-Add telegram link to info page
+-Improve new account creation
+-Improve QR scanner
+-Fix send token issue when user has no tokens
+
+
 ## 0.1.1
 
-Replace 'MetaMask' strings w/ 'Badger'
+-Replace 'MetaMask' strings w/ 'Badger'